+++ src/test/suite/services/bazel-parser.test.ts
@@ -1,7 +1,7 @@
 ////////////////////////////////////////////////////////////////////////////////////
 // MIT License
 //
-// Copyright (c) 2021-2024 NVIDIA Corporation
+// Copyright (c) 2021-2025 NVIDIA Corporation
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -22,6 +22,7 @@
 // SOFTWARE.
 ////////////////////////////////////////////////////////////////////////////////////
 import { BAZEL_BIN, BazelParser } from '../../../services/bazel-parser';
+import { Console } from '../../../services/console';
 import * as assert from 'assert';
 import * as fsPromises from 'fs/promises';
 import * as os from 'os';
@@ -62,6 +63,18 @@ suite('BazelParser', () => {
     let tempDir: string;
 
     setup(async () => {
+        // Initialize Console singleton for tests
+        const mockContext = {
+            extension: {
+                id: 'test.bluebazel',
+                extensionUri: vscode.Uri.file('/test'),
+                packageJSON: {
+                    displayName: 'Blue Bazel Test'
+                }
+            }
+        } as unknown as vscode.ExtensionContext;
+        Console.initialize(mockContext);
+        
         tempDir = await createTempDir();
     });