+++ ReleaseNotes.md
@@ -71,3 +71,12 @@ Fix argument passing when debugging with C/C++ and GDB.
 
 Fix gdbserver not following child processes in C/C++ applications.
 Add support for bzlmod files.
+
+## 1.0.7
+
+Fix typo when running targets.
+Fix shell newlines for the picker.
+Fix tooltips on user buttons.
+Add support for nested Pick and Input with caching (in case the same command appears twice).
+Add support for named problem matchers in custom buttons.
+